Fix "half past" issue in the english translation
Make it correctly show :30 as "half past {current hour}" instead of
"half {next hour}". This also requires some additional checking on the
currently used language (and when to "fall over" to the next hour).

Additionally, express :25 and :35 as "twenty-five past/to", as the form
"five to/past half past ..." is more of a german/alemannic thing.